qgsgrassedit.h:
  - added include for <vector>

qgis.h:

  - added WKBUnknown to enumerated list to be used as error state

qgisapp.cpp:

  - fixed a for loop

qgisapp.h, qgisiface.h, qgslegend.h, qgslegenditem.h, qgsmapcanvs.h, qgsvectorlayer.h:

  - made copy ctor and assignment operator private since we won't be copying
    any of these

qgsattributetable.cpp, qgsrasterlayer.cpp, qgssymbologyutils.cpp:

  - functions that say they return something should actually return something

*renderer.h:

  - made dtors virtual

qgscoordinatetransform.h:

  - added initializer for xMax

qgsgeomtypedialog.cpp:

  - added return of WKBUnknown when given bad data

qgslabel.cpp, qgslabeldialog.cpp:

  - used static_cast to stomp compiler warning

qgsmapcanvas.cpp:

  - made CanvasProperties a class form struct, partly to find work-around for
    gcc bug and partly to ensure that ctor and assignment operator were
    disabled

  - ensured that members were in the same order in initializer list as they
    were declared

  - now use QgsScaleCalculator::units

qgsproject.cpp:

  - now read and write map units; still need to tie it in with more code

qgsprojectproperties.cpp:

  - now use canonical map units stored in QgsProject singleton object

qgsscalecalculator:

  - now prefer to use QgsScaleCalculator::units instead of int


And many minor formatting changes.


git-svn-id: http://svn.osgeo.org/qgis/trunk/qgis@2175 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/qgslabel.cpp
@@ -253,7 +253,7 @@ void QgsLabel::renderLabel( QPainter * painter, QgsRect *viewExtent,
     //
     if (mLabelAttributes->bufferSizeIsSet())
     {
-      int myBufferSize = mLabelAttributes->bufferSize() ;
+      int myBufferSize = static_cast<int>(mLabelAttributes->bufferSize());
       if (mLabelAttributes->bufferColorIsSet())
       {
         painter->setPen( mLabelAttributes->bufferColor());
